Resolve #397, support set style by columns
