+++ excelize_test.go
@@ -244,30 +244,6 @@ func TestNewFile(t *testing.T) {
 	assert.NoError(t, f.SaveAs(filepath.Join("test", "TestNewFile.xlsx")))
 }
 
-func TestColWidth(t *testing.T) {
-	xlsx := NewFile()
-	xlsx.SetColWidth("Sheet1", "B", "A", 12)
-	xlsx.SetColWidth("Sheet1", "A", "B", 12)
-	xlsx.GetColWidth("Sheet1", "A")
-	xlsx.GetColWidth("Sheet1", "C")
-
-	// Test set and get column width with illegal cell coordinates.
-	_, err := xlsx.GetColWidth("Sheet1", "*")
-	assert.EqualError(t, err, `invalid column name "*"`)
-	assert.EqualError(t, xlsx.SetColWidth("Sheet1", "*", "B", 1), `invalid column name "*"`)
-	assert.EqualError(t, xlsx.SetColWidth("Sheet1", "A", "*", 1), `invalid column name "*"`)
-
-	// Test get column width on not exists worksheet.
-	_, err = xlsx.GetColWidth("SheetN", "A")
-	assert.EqualError(t, err, "sheet SheetN is not exist")
-
-	err = xlsx.SaveAs(filepath.Join("test", "TestColWidth.xlsx"))
-	if err != nil {
-		t.Error(err)
-	}
-	convertRowHeightToPixels(0)
-}
-
 func TestAddDrawingVML(t *testing.T) {
 	// Test addDrawingVML with illegal cell coordinates.
 	f := NewFile()
@@ -744,44 +720,6 @@ func TestSheetVisibility(t *testing.T) {
 	assert.NoError(t, f.SaveAs(filepath.Join("test", "TestSheetVisibility.xlsx")))
 }
 
-func TestColumnVisibility(t *testing.T) {
-	t.Run("TestBook1", func(t *testing.T) {
-		f, err := prepareTestBook1()
-		if !assert.NoError(t, err) {
-			t.FailNow()
-		}
-
-		assert.NoError(t, f.SetColVisible("Sheet1", "F", false))
-		assert.NoError(t, f.SetColVisible("Sheet1", "F", true))
-		visible, err := f.GetColVisible("Sheet1", "F")
-		assert.Equal(t, true, visible)
-		assert.NoError(t, err)
-
-		// Test get column visiable on not exists worksheet.
-		_, err = f.GetColVisible("SheetN", "F")
-		assert.EqualError(t, err, "sheet SheetN is not exist")
-
-		// Test get column visiable with illegal cell coordinates.
-		_, err = f.GetColVisible("Sheet1", "*")
-		assert.EqualError(t, err, `invalid column name "*"`)
-		assert.EqualError(t, f.SetColVisible("Sheet1", "*", false), `invalid column name "*"`)
-
-		f.NewSheet("Sheet3")
-		assert.NoError(t, f.SetColVisible("Sheet3", "E", false))
-
-		assert.EqualError(t, f.SetColVisible("SheetN", "E", false), "sheet SheetN is not exist")
-		assert.NoError(t, f.SaveAs(filepath.Join("test", "TestColumnVisibility.xlsx")))
-	})
-
-	t.Run("TestBook3", func(t *testing.T) {
-		f, err := prepareTestBook3()
-		if !assert.NoError(t, err) {
-			t.FailNow()
-		}
-		f.GetColVisible("Sheet1", "B")
-	})
-}
-
 func TestCopySheet(t *testing.T) {
 	f, err := prepareTestBook1()
 	if !assert.NoError(t, err) {
@@ -870,8 +808,11 @@ func TestAddComments(t *testing.T) {
 	}
 
 	s := strings.Repeat("c", 32768)
-	f.AddComment("Sheet1", "A30", `{"author":"`+s+`","text":"`+s+`"}`)
-	f.AddComment("Sheet2", "B7", `{"author":"Excelize: ","text":"This is a comment."}`)
+	assert.NoError(t, f.AddComment("Sheet1", "A30", `{"author":"`+s+`","text":"`+s+`"}`))
+	assert.NoError(t, f.AddComment("Sheet2", "B7", `{"author":"Excelize: ","text":"This is a comment."}`))
+
+	// Test add comment on not exists worksheet.
+	assert.EqualError(t, f.AddComment("SheetN", "B7", `{"author":"Excelize: ","text":"This is a comment."}`), "sheet SheetN is not exist")
 
 	if assert.NoError(t, f.SaveAs(filepath.Join("test", "TestAddComments.xlsx"))) {
 		assert.Len(t, f.GetComments(), 2)
@@ -988,52 +929,6 @@ func TestAddChart(t *testing.T) {
 	assert.NoError(t, f.AddChart("Sheet2", "AN32", `{"type":"area3DPercentStacked","series":[{"name":"Sheet1!$A$30","categories":"Sheet1!$B$29:$D$29","values":"Sheet1!$B$30:$D$30"},{"name":"Sheet1!$A$31","categories":"Sheet1!$B$29:$D$29","values":"Sheet1!$B$31:$D$31"},{"name":"Sheet1!$A$32","categories":"Sheet1!$B$29:$D$29","values":"Sheet1!$B$32:$D$32"}],"format":{"x_scale":1.0,"y_scale":1.0,"x_offset":15,"y_offset":10,"print_obj":true,"lock_aspect_ratio":false,"locked":false},"legend":{"position":"left","show_legend_key":false},"title":{"name":"Fruit 3D 100% Stacked Area Chart"},"plotarea":{"show_bubble_size":true,"show_cat_name":false,"show_leader_lines":false,"show_percent":true,"show_series_name":true,"show_val":true},"show_blanks_as":"zero"}`))
 
 	assert.NoError(t, f.SaveAs(filepath.Join("test", "TestAddChart.xlsx")))
-}
-
-func TestInsertCol(t *testing.T) {
-	f := NewFile()
-	sheet1 := f.GetSheetName(1)
-
-	fillCells(f, sheet1, 10, 10)
-
-	f.SetCellHyperLink(sheet1, "A5", "https://github.com/360EntSecGroup-Skylar/excelize", "External")
-	f.MergeCell(sheet1, "A1", "C3")
-
-	err := f.AutoFilter(sheet1, "A2", "B2", `{"column":"B","expression":"x != blanks"}`)
-	if !assert.NoError(t, err) {
-		t.FailNow()
-	}
-
-	assert.NoError(t, f.InsertCol(sheet1, "A"))
-
-	// Test insert column with illegal cell coordinates.
-	assert.EqualError(t, f.InsertCol("Sheet1", "*"), `invalid column name "*"`)
-
-	assert.NoError(t, f.SaveAs(filepath.Join("test", "TestInsertCol.xlsx")))
-}
-
-func TestRemoveCol(t *testing.T) {
-	f := NewFile()
-	sheet1 := f.GetSheetName(1)
-
-	fillCells(f, sheet1, 10, 15)
-
-	f.SetCellHyperLink(sheet1, "A5", "https://github.com/360EntSecGroup-Skylar/excelize", "External")
-	f.SetCellHyperLink(sheet1, "C5", "https://github.com", "External")
-
-	f.MergeCell(sheet1, "A1", "B1")
-	f.MergeCell(sheet1, "A2", "B2")
-
-	assert.NoError(t, f.RemoveCol(sheet1, "A"))
-	assert.NoError(t, f.RemoveCol(sheet1, "A"))
-
-	// Test remove column with illegal cell coordinates.
-	assert.EqualError(t, f.RemoveCol("Sheet1", "*"), `invalid column name "*"`)
-
-	// Test remove column on not exists worksheet.
-	assert.EqualError(t, f.RemoveCol("SheetN", "B"), "sheet SheetN is not exist")
-
-	assert.NoError(t, f.SaveAs(filepath.Join("test", "TestRemoveCol.xlsx")))
 }
 
 func TestSetPane(t *testing.T) {
@@ -1160,48 +1055,6 @@ func TestSetSheetRow(t *testing.T) {
 	assert.EqualError(t, f.SetSheetRow("Sheet1", "B27", []interface{}{}), `pointer to slice expected`)
 	assert.EqualError(t, f.SetSheetRow("Sheet1", "B27", &f), `pointer to slice expected`)
 	assert.NoError(t, f.SaveAs(filepath.Join("test", "TestSetSheetRow.xlsx")))
-}
-
-func TestOutlineLevel(t *testing.T) {
-	f := NewFile()
-	f.NewSheet("Sheet2")
-	f.SetColOutlineLevel("Sheet1", "D", 4)
-	f.GetColOutlineLevel("Sheet1", "D")
-	f.GetColOutlineLevel("Shee2", "A")
-	f.SetColWidth("Sheet2", "A", "D", 13)
-	f.SetColOutlineLevel("Sheet2", "B", 2)
-	f.SetRowOutlineLevel("Sheet1", 2, 250)
-
-	// Test set and get column outline level with illegal cell coordinates.
-	assert.EqualError(t, f.SetColOutlineLevel("Sheet1", "*", 1), `invalid column name "*"`)
-	_, err := f.GetColOutlineLevel("Sheet1", "*")
-	assert.EqualError(t, err, `invalid column name "*"`)
-
-	// Test set column outline level on not exists worksheet.
-	assert.EqualError(t, f.SetColOutlineLevel("SheetN", "E", 2), "sheet SheetN is not exist")
-
-	assert.EqualError(t, f.SetRowOutlineLevel("Sheet1", 0, 1), "invalid row number 0")
-	level, err := f.GetRowOutlineLevel("Sheet1", 2)
-	assert.NoError(t, err)
-	assert.Equal(t, uint8(250), level)
-
-	_, err = f.GetRowOutlineLevel("Sheet1", 0)
-	assert.EqualError(t, err, `invalid row number 0`)
-
-	level, err = f.GetRowOutlineLevel("Sheet1", 10)
-	assert.NoError(t, err)
-	assert.Equal(t, uint8(0), level)
-
-	err = f.SaveAs(filepath.Join("test", "TestOutlineLevel.xlsx"))
-	if !assert.NoError(t, err) {
-		t.FailNow()
-	}
-
-	f, err = OpenFile(filepath.Join("test", "Book1.xlsx"))
-	if !assert.NoError(t, err) {
-		t.FailNow()
-	}
-	f.SetColOutlineLevel("Sheet2", "B", 2)
 }
 
 func TestThemeColor(t *testing.T) {
